+++ doc/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+2.3.0 - released 2025-02-09
+---------------------------
+
+* New: add :func:`wasabigeom.quadratic_bezier` and
+  :func:`wasabigeom.cubic_bezier` to subdivide quadratic and cubic Bézier curves
+  for rasterisation.
+
+
 2.2.0 - released 2024-04-01
 ---------------------------